requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-decouple==3.8
openpyxl==3.1.2
pillow==10.1.0
aiofiles==23.2.1
pydantic[email]